UNIQUE BUSINESS STRATEGIES EFFECTIVENESS OF UNIQUE TAGS AND GIMMICKS AS KEYS TO ACHIEVE SATISFACTION
Abstract

A big percentage of customers are always looking for eateries that can satisfy their needs which can catch their attention and something that is new to their taste. They crave for something that can reach their standards when it comes to food. The main objective of this quantitative study is to identify the effectiveness of unusual business strategies in attaining customers’ satisfaction by measuring profit. 60 respondents with 21 sample respondents from the business simulation were selected. The type of hypothesis used in this study was null hypothesis. True experimental and evaluative approaches were used as the most appropriate research designs of the study. The researchers used t test and frequency distribution to process the data. Pre – test and post – test and survey questionnaire were used to obtain additional data needed. The null hypothesis of the study was accepted. Therefore, there is no significant difference in the profitability of applying unusual business strategies to a normal business technique. Moreover, it was shown in the findings of the study that using unique strategies has no effect to the profit of business. The researchers recommend that business owners may add cost to the products to increase profit. Furthermore, through survey, it was revealed that customers were positively affected by the unique business strategies used.