Baetiong, Lance Jhorick S. Lunes, Elvie L. Montoya, Jan Jerwin C. Robles, Mary Rose Joy L. Tanig, Andrei France S.
Abstract
Business is an institution which produce goods and services demanded by people. As every
human being has needs to be attained to survive, this study aims to help the community of Tibaguin
Hagonoy, Bulacan which is a coastal area. This research focuses on identifying the feasibility of
businesses that can be established in Tibaguin. Based on the findings, 53.75% out of 100% were in
favor of establishing a pharmacy in their barangay. Then, the researchers conducted survey for 100
respondents that lives in Tibaguin to test if the business is feasible to establish in their community. The
results revealed that 100% of the respondents answered that they are in favor for establishing pharmacy
in Tibaguin. As a result, the researchers found out that the business is feasible for establishment in
Tibaguin Hagonoy, Bulacan.